r/aws_cloudwatch_event_rule: Migrate to AWS SDK for Go v2.
