+++ internal/service/events/id_test.go
@@ -9,130 +9,6 @@ import (
 	tfevents "github.com/hashicorp/terraform-provider-aws/internal/service/events"
 )
 
-func TestRuleParseResourceID(t *testing.T) {
-	t.Parallel()
-
-	testCases := []struct {
-		TestName      string
-		InputID       string
-		ExpectedError bool
-		ExpectedPart0 string
-		ExpectedPart1 string
-	}{
-		{
-			TestName:      "empty ID",
-			InputID:       "",
-			ExpectedError: true,
-		},
-		{
-			TestName:      "single part",
-			InputID:       "TestRule",
-			ExpectedPart0: tfevents.DefaultEventBusName,
-			ExpectedPart1: "TestRule",
-		},
-		{
-			TestName:      "two parts",
-			InputID:       tfevents.RuleCreateResourceID("TestEventBus", "TestRule"),
-			ExpectedPart0: "TestEventBus",
-			ExpectedPart1: "TestRule",
-		},
-		{
-			TestName:      "two parts with default event bus",
-			InputID:       tfevents.RuleCreateResourceID(tfevents.DefaultEventBusName, "TestRule"),
-			ExpectedPart0: tfevents.DefaultEventBusName,
-			ExpectedPart1: "TestRule",
-		},
-		{
-			TestName:      "partner event bus 1",
-			InputID:       "aws.partner/example.com/Test/TestRule",
-			ExpectedPart0: "aws.partner/example.com/Test",
-			ExpectedPart1: "TestRule",
-		},
-		{
-			TestName:      "partner event bus 2",
-			InputID:       "aws.partner/example.net/id/18554d09-58ff-aa42-ba9c-c4c33899006f/test",
-			ExpectedPart0: "aws.partner/example.net/id/18554d09-58ff-aa42-ba9c-c4c33899006f",
-			ExpectedPart1: "test",
-		},
-		{
-			TestName: "ARN event bus",
-			//lintignore:AWSAT003,AWSAT005
-			InputID: tfevents.RuleCreateResourceID("arn:aws:events:us-east-2:123456789012:event-bus/default", "TestRule"),
-			//lintignore:AWSAT003,AWSAT005
-			ExpectedPart0: "arn:aws:events:us-east-2:123456789012:event-bus/default",
-			ExpectedPart1: "TestRule",
-		},
-		{
-			TestName: "ARN based partner event bus",
-			// lintignore:AWSAT003,AWSAT005
-			InputID: "arn:aws:events:us-east-2:123456789012:event-bus/aws.partner/genesys.com/cloud/a12bc345-d678-90e1-2f34-gh5678i9012ej/_genesys/TestRule",
-			// lintignore:AWSAT003,AWSAT005
-			ExpectedPart0: "arn:aws:events:us-east-2:123456789012:event-bus/aws.partner/genesys.com/cloud/a12bc345-d678-90e1-2f34-gh5678i9012ej/_genesys",
-			ExpectedPart1: "TestRule",
-		},
-		{
-			TestName:      "empty both parts",
-			InputID:       "/",
-			ExpectedError: true,
-		},
-		{
-			TestName:      "empty first part",
-			InputID:       "/TestRule",
-			ExpectedError: true,
-		},
-		{
-			TestName:      "empty second part",
-			InputID:       "TestEventBus/",
-			ExpectedError: true,
-		},
-		{
-			TestName:      "empty partner event rule",
-			InputID:       "aws.partner/example.com/Test/",
-			ExpectedError: true,
-		},
-		{
-			TestName:      "three parts",
-			InputID:       "TestEventBus/TestRule/Suffix",
-			ExpectedError: true,
-		},
-		{
-			TestName:      "four parts",
-			InputID:       "abc.partner/TestEventBus/TestRule/Suffix",
-			ExpectedError: true,
-		},
-		{
-			TestName:      "five parts",
-			InputID:       "test/aws.partner/example.com/Test/TestRule",
-			ExpectedError: true,
-		},
-	}
-
-	for _, testCase := range testCases {
-		testCase := testCase
-		t.Run(testCase.TestName, func(t *testing.T) {
-			t.Parallel()
-
-			gotPart0, gotPart1, err := tfevents.RuleParseResourceID(testCase.InputID)
-
-			if err == nil && testCase.ExpectedError {
-				t.Fatalf("expected error, got no error")
-			}
-
-			if err != nil && !testCase.ExpectedError {
-				t.Fatalf("got unexpected error: %s", err)
-			}
-
-			if gotPart0 != testCase.ExpectedPart0 {
-				t.Errorf("got part 0 %s, expected %s", gotPart0, testCase.ExpectedPart0)
-			}
-
-			if gotPart1 != testCase.ExpectedPart1 {
-				t.Errorf("got part 1 %s, expected %s", gotPart1, testCase.ExpectedPart1)
-			}
-		})
-	}
-}
-
 func TestTargetParseImportID(t *testing.T) {
 	t.Parallel()
