+++ internal/service/events/rule.go
@@ -9,17 +9,20 @@ import (
 	"encoding/json"
 	"fmt"
 	"log"
+	"strings"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/eventbridge"
-	"github.com/hashicorp/aws-sdk-go-base/v2/awsv1shim/v2/tfawserr"
+	"github.com/YakDriver/regexache"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
+	"github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
 	"github.com/hashicorp/terraform-provider-aws/internal/create"
+	"github.com/hashicorp/terraform-provider-aws/internal/enum"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs/sdkdiag"
 	tftags "github.com/hashicorp/terraform-provider-aws/internal/tags"
@@ -28,14 +31,9 @@ import (
 	"github.com/hashicorp/terraform-provider-aws/names"
 )
 
-const (
-	ruleCreateRetryTimeout = 2 * time.Minute
-	ruleDeleteRetryTimeout = 5 * time.Minute
-)
-
 // @SDKResource("aws_cloudwatch_event_rule", name="Rule")
 // @Tags(identifierAttribute="arn")
-func ResourceRule() *schema.Resource {
+func resourceRule() *schema.Resource {
 	return &schema.Resource{
 		CreateWithoutTimeout: resourceRuleCreate,
 		ReadWithoutTimeout:   resourceRuleRead,
@@ -78,7 +76,7 @@ func ResourceRule() *schema.Resource {
 				ValidateFunc: validateEventPatternValue(),
 				AtLeastOneOf: []string{"schedule_expression", "event_pattern"},
 				StateFunc: func(v interface{}) string {
-					json, _ := RuleEventPatternJSONDecoder(v.(string))
+					json, _ := ruleEventPatternJSONDecoder(v.(string))
 					return json
 				},
 			},
@@ -123,12 +121,9 @@ func ResourceRule() *schema.Resource {
 				AtLeastOneOf: []string{"schedule_expression", "event_pattern"},
 			},
 			"state": {
-				Type:     schema.TypeString,
-				Optional: true,
-				ValidateFunc: validation.StringInSlice(
-					eventbridge.RuleState_Values(),
-					false,
-				),
+				Type:             schema.TypeString,
+				Optional:         true,
+				ValidateDiagFunc: enum.Validate[types.RuleState](),
 				DiffSuppressFunc: func(k, oldValue, newValue string, d *schema.ResourceData) bool {
 					if oldValue != "" && newValue == "" {
 						return true
@@ -149,7 +144,7 @@ func ResourceRule() *schema.Resource {
 
 func resourceRuleCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
-	conn := meta.(*conns.AWSClient).EventsConn(ctx)
+	conn := meta.(*conns.AWSClient).EventsClient(ctx)
 
 	name := create.Name(d.Get("name").(string), d.Get("name_prefix").(string))
 	input := expandPutRuleInput(d, name)
@@ -158,7 +153,7 @@ func resourceRuleCreate(ctx context.Context, d *schema.ResourceData, meta interf
 	arn, err := retryPutRule(ctx, conn, input)
 
 	// Some partitions (e.g. ISO) may not support tag-on-create.
-	if input.Tags != nil && errs.IsUnsupportedOperationInPartitionError(conn.PartitionID, err) {
+	if input.Tags != nil && errs.IsUnsupportedOperationInPartitionError(meta.(*conns.AWSClient).Partition, err) {
 		input.Tags = nil
 
 		arn, err = retryPutRule(ctx, conn, input)
@@ -168,11 +163,14 @@ func resourceRuleCreate(ctx context.Context, d *schema.ResourceData, meta interf
 		return sdkdiag.AppendErrorf(diags, "creating EventBridge Rule (%s): %s", name, err)
 	}
 
-	eventBusName, ruleName := aws.StringValue(input.EventBusName), aws.StringValue(input.Name)
-	d.SetId(RuleCreateResourceID(eventBusName, ruleName))
+	eventBusName, ruleName := aws.ToString(input.EventBusName), aws.ToString(input.Name)
+	d.SetId(ruleCreateResourceID(eventBusName, ruleName))
 
-	_, err = tfresource.RetryWhenNotFound(ctx, ruleCreateRetryTimeout, func() (interface{}, error) {
-		return FindRuleByTwoPartKey(ctx, conn, eventBusName, ruleName)
+	const (
+		timeout = 2 * time.Minute
+	)
+	_, err = tfresource.RetryWhenNotFound(ctx, timeout, func() (interface{}, error) {
+		return findRuleByTwoPartKey(ctx, conn, eventBusName, ruleName)
 	})
 
 	if err != nil {
@@ -184,7 +182,7 @@ func resourceRuleCreate(ctx context.Context, d *schema.ResourceData, meta interf
 		err := createTags(ctx, conn, arn, tags)
 
 		// If default tags only, continue. Otherwise, error.
-		if v, ok := d.GetOk(names.AttrTags); (!ok || len(v.(map[string]interface{})) == 0) && errs.IsUnsupportedOperationInPartitionError(conn.PartitionID, err) {
+		if v, ok := d.GetOk(names.AttrTags); (!ok || len(v.(map[string]interface{})) == 0) && errs.IsUnsupportedOperationInPartitionError(meta.(*conns.AWSClient).Partition, err) {
 			return append(diags, resourceRuleRead(ctx, d, meta)...)
 		}
 
@@ -198,15 +196,14 @@ func resourceRuleCreate(ctx context.Context, d *schema.ResourceData, meta interf
 
 func resourceRuleRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
-	conn := meta.(*conns.AWSClient).EventsConn(ctx)
+	conn := meta.(*conns.AWSClient).EventsClient(ctx)
 
-	eventBusName, ruleName, err := RuleParseResourceID(d.Id())
-
+	eventBusName, ruleName, err := ruleParseResourceID(d.Id())
 	if err != nil {
 		return sdkdiag.AppendFromErr(diags, err)
 	}
 
-	output, err := FindRuleByTwoPartKey(ctx, conn, eventBusName, ruleName)
+	output, err := findRuleByTwoPartKey(ctx, conn, eventBusName, ruleName)
 
 	if !d.IsNewResource() && tfresource.NotFound(err) {
 		log.Printf("[WARN] EventBridge Rule (%s) not found, removing from state", d.Id())
@@ -218,45 +215,44 @@ func resourceRuleRead(ctx context.Context, d *schema.ResourceData, meta interfac
 		return sdkdiag.AppendErrorf(diags, "reading EventBridge Rule (%s): %s", d.Id(), err)
 	}
 
-	arn := aws.StringValue(output.Arn)
+	arn := aws.ToString(output.Arn)
 	d.Set("arn", arn)
 	d.Set("description", output.Description)
 	d.Set("event_bus_name", eventBusName) // Use event bus name from resource ID as API response may collapse any ARN.
 	if output.EventPattern != nil {
-		pattern, err := RuleEventPatternJSONDecoder(aws.StringValue(output.EventPattern))
+		pattern, err := ruleEventPatternJSONDecoder(aws.ToString(output.EventPattern))
 		if err != nil {
-			return sdkdiag.AppendErrorf(diags, "event pattern contains an invalid JSON: %s", err)
+			return sdkdiag.AppendFromErr(diags, err)
 		}
 		d.Set("event_pattern", pattern)
 	}
-	switch aws.StringValue(output.State) {
-	case eventbridge.RuleStateEnabled,
-		eventbridge.RuleStateEnabledWithAllCloudtrailManagementEvents:
+	switch output.State {
+	case types.RuleStateEnabled, types.RuleStateEnabledWithAllCloudtrailManagementEvents:
 		d.Set("is_enabled", true)
 	default:
 		d.Set("is_enabled", false)
 	}
-	d.Set("state", output.State)
 	d.Set("name", output.Name)
-	d.Set("name_prefix", create.NamePrefixFromName(aws.StringValue(output.Name)))
+	d.Set("name_prefix", create.NamePrefixFromName(aws.ToString(output.Name)))
 	d.Set("role_arn", output.RoleArn)
 	d.Set("schedule_expression", output.ScheduleExpression)
+	d.Set("state", output.State)
 
 	return diags
 }
 
 func resourceRuleUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
-	conn := meta.(*conns.AWSClient).EventsConn(ctx)
+	conn := meta.(*conns.AWSClient).EventsClient(ctx)
 
 	if d.HasChangesExcept("tags", "tags_all") {
-		_, ruleName, err := RuleParseResourceID(d.Id())
-
+		_, ruleName, err := ruleParseResourceID(d.Id())
 		if err != nil {
 			return sdkdiag.AppendFromErr(diags, err)
 		}
 
 		input := expandPutRuleInput(d, ruleName)
+
 		_, err = retryPutRule(ctx, conn, input)
 
 		if err != nil {
@@ -269,10 +265,9 @@ func resourceRuleUpdate(ctx context.Context, d *schema.ResourceData, meta interf
 
 func resourceRuleDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
-	conn := meta.(*conns.AWSClient).EventsConn(ctx)
+	conn := meta.(*conns.AWSClient).EventsClient(ctx)
 
-	eventBusName, ruleName, err := RuleParseResourceID(d.Id())
-
+	eventBusName, ruleName, err := ruleParseResourceID(d.Id())
 	if err != nil {
 		return sdkdiag.AppendFromErr(diags, err)
 	}
@@ -284,12 +279,15 @@ func resourceRuleDelete(ctx context.Context, d *schema.ResourceData, meta interf
 		input.EventBusName = aws.String(eventBusName)
 	}
 
+	const (
+		timeout = 5 * time.Minute
+	)
 	log.Printf("[DEBUG] Deleting EventBridge Rule: %s", d.Id())
-	_, err = tfresource.RetryWhenAWSErrMessageContains(ctx, ruleDeleteRetryTimeout, func() (interface{}, error) {
-		return conn.DeleteRuleWithContext(ctx, input)
-	}, "ValidationException", "Rule can't be deleted since it has targets")
+	_, err = tfresource.RetryWhenAWSErrMessageContains(ctx, timeout, func() (interface{}, error) {
+		return conn.DeleteRule(ctx, input)
+	}, errCodeValidationException, "Rule can't be deleted since it has targets")
 
-	if tfawserr.ErrCodeEquals(err, eventbridge.ErrCodeResourceNotFoundException) {
+	if errs.IsA[*types.ResourceNotFoundException](err) {
 		return diags
 	}
 
@@ -300,29 +298,29 @@ func resourceRuleDelete(ctx context.Context, d *schema.ResourceData, meta interf
 	return diags
 }
 
-func retryPutRule(ctx context.Context, conn *eventbridge.EventBridge, input *eventbridge.PutRuleInput) (string, error) {
+func retryPutRule(ctx context.Context, conn *eventbridge.Client, input *eventbridge.PutRuleInput) (string, error) {
 	outputRaw, err := tfresource.RetryWhenAWSErrMessageContains(ctx, propagationTimeout, func() (interface{}, error) {
-		return conn.PutRuleWithContext(ctx, input)
-	}, "ValidationException", "cannot be assumed by principal")
+		return conn.PutRule(ctx, input)
+	}, errCodeValidationException, "cannot be assumed by principal")
 
 	if err != nil {
 		return "", err
 	}
 
-	return aws.StringValue(outputRaw.(*eventbridge.PutRuleOutput).RuleArn), nil
+	return aws.ToString(outputRaw.(*eventbridge.PutRuleOutput).RuleArn), nil
 }
 
-func FindRuleByTwoPartKey(ctx context.Context, conn *eventbridge.EventBridge, eventBusName, ruleName string) (*eventbridge.DescribeRuleOutput, error) {
-	input := eventbridge.DescribeRuleInput{
+func findRuleByTwoPartKey(ctx context.Context, conn *eventbridge.Client, eventBusName, ruleName string) (*eventbridge.DescribeRuleOutput, error) {
+	input := &eventbridge.DescribeRuleInput{
 		Name: aws.String(ruleName),
 	}
 	if eventBusName != "" {
 		input.EventBusName = aws.String(eventBusName)
 	}
 
-	output, err := conn.DescribeRuleWithContext(ctx, &input)
+	output, err := conn.DescribeRule(ctx, input)
 
-	if tfawserr.ErrCodeEquals(err, eventbridge.ErrCodeResourceNotFoundException) {
+	if errs.IsA[*types.ResourceNotFoundException](err) {
 		return nil, &retry.NotFoundError{
 			LastError:   err,
 			LastRequest: input,
@@ -340,8 +338,50 @@ func FindRuleByTwoPartKey(ctx context.Context, conn *eventbridge.EventBridge, ev
 	return output, nil
 }
 
-// RuleEventPatternJSONDecoder decodes unicode translation of <,>,&
-func RuleEventPatternJSONDecoder(jsonString interface{}) (string, error) {
+var (
+	eventBusARNPattern     = regexache.MustCompile(`^arn:aws[\w-]*:events:[a-z]{2}-[a-z]+-[\w-]+:[0-9]{12}:event-bus\/[0-9A-Za-z_.-]+$`)
+	partnerEventBusPattern = regexache.MustCompile(`^(?:arn:aws[\w-]*:events:[a-z]{2}-[a-z]+-[\w-]+:[0-9]{12}:event-bus\/)?aws\.partner(/[0-9A-Za-z_.-]+){2,}$`)
+)
+
+const ruleResourceIDSeparator = "/"
+
+func ruleCreateResourceID(eventBusName, ruleName string) string {
+	if eventBusName == "" || eventBusName == DefaultEventBusName {
+		return ruleName
+	}
+
+	parts := []string{eventBusName, ruleName}
+	id := strings.Join(parts, ruleResourceIDSeparator)
+
+	return id
+}
+
+func ruleParseResourceID(id string) (string, string, error) {
+	parts := strings.Split(id, ruleResourceIDSeparator)
+
+	if len(parts) == 1 && parts[0] != "" {
+		return DefaultEventBusName, parts[0], nil
+	}
+	if len(parts) == 2 && parts[0] != "" && parts[1] != "" {
+		return parts[0], parts[1], nil
+	}
+	if len(parts) > 2 {
+		i := strings.LastIndex(id, ruleResourceIDSeparator)
+		eventBusName := id[:i]
+		ruleName := id[i+1:]
+		if eventBusARNPattern.MatchString(eventBusName) && ruleName != "" {
+			return eventBusName, ruleName, nil
+		}
+		if partnerEventBusPattern.MatchString(eventBusName) && ruleName != "" {
+			return eventBusName, ruleName, nil
+		}
+	}
+
+	return "", "", fmt.Errorf("unexpected format for ID (%[1]s), expected EVENTBUSNAME%[2]sRULENAME or RULENAME", id, ruleResourceIDSeparator)
+}
+
+// ruleEventPatternJSONDecoder decodes unicode translation of <,>,&
+func ruleEventPatternJSONDecoder(jsonString interface{}) (string, error) {
 	var j interface{}
 
 	if jsonString == nil || jsonString.(string) == "" {
@@ -382,7 +422,7 @@ func expandPutRuleInput(d *schema.ResourceData, name string) *eventbridge.PutRul
 	}
 
 	if v, ok := d.GetOk("event_pattern"); ok {
-		json, _ := RuleEventPatternJSONDecoder(v.(string))
+		json, _ := ruleEventPatternJSONDecoder(v.(string))
 		apiObject.EventPattern = aws.String(json)
 	}
 
@@ -397,14 +437,14 @@ func expandPutRuleInput(d *schema.ResourceData, name string) *eventbridge.PutRul
 	rawConfig := d.GetRawConfig()
 	rawState := rawConfig.GetAttr("state")
 	if rawState.IsKnown() && !rawState.IsNull() {
-		apiObject.State = aws.String(rawState.AsString())
+		apiObject.State = types.RuleState(rawState.AsString())
 	} else {
 		rawIsEnabled := rawConfig.GetAttr("is_enabled")
 		if rawIsEnabled.IsKnown() && !rawIsEnabled.IsNull() {
 			if rawIsEnabled.True() {
-				apiObject.State = aws.String(eventbridge.RuleStateEnabled)
+				apiObject.State = types.RuleStateEnabled
 			} else {
-				apiObject.State = aws.String(eventbridge.RuleStateDisabled)
+				apiObject.State = types.RuleStateDisabled
 			}
 		}
 	}
@@ -414,7 +454,7 @@ func expandPutRuleInput(d *schema.ResourceData, name string) *eventbridge.PutRul
 
 func validateEventPatternValue() schema.SchemaValidateFunc {
 	return func(v interface{}, k string) (ws []string, errors []error) {
-		json, err := RuleEventPatternJSONDecoder(v.(string))
+		json, err := ruleEventPatternJSONDecoder(v.(string))
 		if err != nil {
 			errors = append(errors, fmt.Errorf("%q contains an invalid JSON: %w", k, err))
