+++ internal/service/events/rule_migrate.go
@@ -6,7 +6,7 @@ package events
 import (
 	"context"
 
-	"github.com/aws/aws-sdk-go/service/eventbridge"
+	"github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	tftags "github.com/hashicorp/terraform-provider-aws/internal/tags"
@@ -35,7 +35,7 @@ func resourceRuleV0() *schema.Resource {
 				Type:     schema.TypeString,
 				Optional: true,
 				StateFunc: func(v interface{}) string {
-					json, _ := RuleEventPatternJSONDecoder(v.(string))
+					json, _ := ruleEventPatternJSONDecoder(v.(string))
 					return json
 				},
 			},
@@ -85,9 +85,9 @@ func resourceRuleUpgradeV0(ctx context.Context, rawState map[string]any, meta an
 	})
 
 	if rawState["is_enabled"].(bool) {
-		rawState["state"] = eventbridge.RuleStateEnabled
+		rawState["state"] = types.RuleStateEnabled
 	} else {
-		rawState["state"] = eventbridge.RuleStateDisabled
+		rawState["state"] = types.RuleStateDisabled
 	}
 
 	return rawState, nil
