+++ internal/service/events/rule_test.go
@@ -10,8 +10,9 @@ import (
 	"testing"
 
 	"github.com/YakDriver/regexache"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/eventbridge"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
+	"github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
 	"github.com/google/go-cmp/cmp"
 	sdkacctest "github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
@@ -33,6 +34,130 @@ func testAccErrorCheckSkip(t *testing.T) resource.ErrorCheckFunc {
 		"Operation is disabled in this region",
 		"not a supported service for a target",
 	)
+}
+
+func TestRuleParseResourceID(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		TestName      string
+		InputID       string
+		ExpectedError bool
+		ExpectedPart0 string
+		ExpectedPart1 string
+	}{
+		{
+			TestName:      "empty ID",
+			InputID:       "",
+			ExpectedError: true,
+		},
+		{
+			TestName:      "single part",
+			InputID:       "TestRule",
+			ExpectedPart0: tfevents.DefaultEventBusName,
+			ExpectedPart1: "TestRule",
+		},
+		{
+			TestName:      "two parts",
+			InputID:       tfevents.RuleCreateResourceID("TestEventBus", "TestRule"),
+			ExpectedPart0: "TestEventBus",
+			ExpectedPart1: "TestRule",
+		},
+		{
+			TestName:      "two parts with default event bus",
+			InputID:       tfevents.RuleCreateResourceID(tfevents.DefaultEventBusName, "TestRule"),
+			ExpectedPart0: tfevents.DefaultEventBusName,
+			ExpectedPart1: "TestRule",
+		},
+		{
+			TestName:      "partner event bus 1",
+			InputID:       "aws.partner/example.com/Test/TestRule",
+			ExpectedPart0: "aws.partner/example.com/Test",
+			ExpectedPart1: "TestRule",
+		},
+		{
+			TestName:      "partner event bus 2",
+			InputID:       "aws.partner/example.net/id/18554d09-58ff-aa42-ba9c-c4c33899006f/test",
+			ExpectedPart0: "aws.partner/example.net/id/18554d09-58ff-aa42-ba9c-c4c33899006f",
+			ExpectedPart1: "test",
+		},
+		{
+			TestName: "ARN event bus",
+			//lintignore:AWSAT003,AWSAT005
+			InputID: tfevents.RuleCreateResourceID("arn:aws:events:us-east-2:123456789012:event-bus/default", "TestRule"),
+			//lintignore:AWSAT003,AWSAT005
+			ExpectedPart0: "arn:aws:events:us-east-2:123456789012:event-bus/default",
+			ExpectedPart1: "TestRule",
+		},
+		{
+			TestName: "ARN based partner event bus",
+			// lintignore:AWSAT003,AWSAT005
+			InputID: "arn:aws:events:us-east-2:123456789012:event-bus/aws.partner/genesys.com/cloud/a12bc345-d678-90e1-2f34-gh5678i9012ej/_genesys/TestRule",
+			// lintignore:AWSAT003,AWSAT005
+			ExpectedPart0: "arn:aws:events:us-east-2:123456789012:event-bus/aws.partner/genesys.com/cloud/a12bc345-d678-90e1-2f34-gh5678i9012ej/_genesys",
+			ExpectedPart1: "TestRule",
+		},
+		{
+			TestName:      "empty both parts",
+			InputID:       "/",
+			ExpectedError: true,
+		},
+		{
+			TestName:      "empty first part",
+			InputID:       "/TestRule",
+			ExpectedError: true,
+		},
+		{
+			TestName:      "empty second part",
+			InputID:       "TestEventBus/",
+			ExpectedError: true,
+		},
+		{
+			TestName:      "empty partner event rule",
+			InputID:       "aws.partner/example.com/Test/",
+			ExpectedError: true,
+		},
+		{
+			TestName:      "three parts",
+			InputID:       "TestEventBus/TestRule/Suffix",
+			ExpectedError: true,
+		},
+		{
+			TestName:      "four parts",
+			InputID:       "abc.partner/TestEventBus/TestRule/Suffix",
+			ExpectedError: true,
+		},
+		{
+			TestName:      "five parts",
+			InputID:       "test/aws.partner/example.com/Test/TestRule",
+			ExpectedError: true,
+		},
+	}
+
+	for _, testCase := range testCases {
+		testCase := testCase
+		t.Run(testCase.TestName, func(t *testing.T) {
+			t.Parallel()
+
+			gotPart0, gotPart1, err := tfevents.RuleParseResourceID(testCase.InputID)
+
+			if err == nil && testCase.ExpectedError {
+				t.Fatalf("expected error, got no error")
+			}
+
+			if err != nil && !testCase.ExpectedError {
+				t.Fatalf("got unexpected error: %s", err)
+			}
+
+			if gotPart0 != testCase.ExpectedPart0 {
+				t.Errorf("got part 0 %s, expected %s", gotPart0, testCase.ExpectedPart0)
+			}
+
+			if gotPart1 != testCase.ExpectedPart1 {
+				t.Errorf("got part 1 %s, expected %s", gotPart1, testCase.ExpectedPart1)
+			}
+		})
+	}
 }
 
 func TestRuleEventPatternJSONDecoder(t *testing.T) {
@@ -523,12 +648,12 @@ func TestAccEventsRule_state(t *testing.T) {
 		CheckDestroy:             testAccCheckRuleDestroy(ctx),
 		Steps: []resource.TestStep{
 			{
-				Config: testAccRuleConfig_state(rName, eventbridge.RuleStateDisabled),
+				Config: testAccRuleConfig_state(rName, string(types.RuleStateDisabled)),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					testAccCheckRuleExists(ctx, resourceName, &v1),
 					resource.TestCheckResourceAttr(resourceName, "is_enabled", "false"),
-					resource.TestCheckResourceAttr(resourceName, "state", eventbridge.RuleStateDisabled),
-					testAccCheckRuleEnabled(ctx, resourceName, eventbridge.RuleStateDisabled),
+					resource.TestCheckResourceAttr(resourceName, "state", string(types.RuleStateDisabled)),
+					testAccCheckRuleEnabled(ctx, resourceName, types.RuleStateDisabled),
 				),
 			},
 			{
@@ -537,12 +662,12 @@ func TestAccEventsRule_state(t *testing.T) {
 				ImportStateVerify: true,
 			},
 			{
-				Config: testAccRuleConfig_state(rName, eventbridge.RuleStateEnabled),
+				Config: testAccRuleConfig_state(rName, string(types.RuleStateEnabled)),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					testAccCheckRuleExists(ctx, resourceName, &v2),
 					resource.TestCheckResourceAttr(resourceName, "is_enabled", "true"),
-					resource.TestCheckResourceAttr(resourceName, "state", eventbridge.RuleStateEnabled),
-					testAccCheckRuleEnabled(ctx, resourceName, eventbridge.RuleStateEnabled),
+					resource.TestCheckResourceAttr(resourceName, "state", string(types.RuleStateEnabled)),
+					testAccCheckRuleEnabled(ctx, resourceName, types.RuleStateEnabled),
 				),
 			},
 			{
@@ -551,12 +676,12 @@ func TestAccEventsRule_state(t *testing.T) {
 				ImportStateVerify: true,
 			},
 			{
-				Config: testAccRuleConfig_state(rName, eventbridge.RuleStateEnabledWithAllCloudtrailManagementEvents),
+				Config: testAccRuleConfig_state(rName, string(types.RuleStateEnabledWithAllCloudtrailManagementEvents)),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					testAccCheckRuleExists(ctx, resourceName, &v3),
 					resource.TestCheckResourceAttr(resourceName, "is_enabled", "true"),
-					resource.TestCheckResourceAttr(resourceName, "state", eventbridge.RuleStateEnabledWithAllCloudtrailManagementEvents),
-					testAccCheckRuleEnabled(ctx, resourceName, eventbridge.RuleStateEnabledWithAllCloudtrailManagementEvents),
+					resource.TestCheckResourceAttr(resourceName, "state", string(types.RuleStateEnabledWithAllCloudtrailManagementEvents)),
+					testAccCheckRuleEnabled(ctx, resourceName, types.RuleStateEnabledWithAllCloudtrailManagementEvents),
 				),
 			},
 			{
@@ -657,7 +782,7 @@ func TestAccEventsRule_migrateV0(t *testing.T) {
 	testcases := map[string]struct {
 		config            string
 		expectedIsEnabled string
-		expectedState     string
+		expectedState     types.RuleState
 	}{
 		"basic": {
 			config:            testAccRuleConfig_basic(sdkacctest.RandomWithPrefix(acctest.ResourcePrefix)),
@@ -714,7 +839,7 @@ func TestAccEventsRule_migrateV0(t *testing.T) {
 						},
 						Check: resource.ComposeAggregateTestCheckFunc(
 							resource.TestCheckResourceAttr(resourceName, "is_enabled", testcase.expectedIsEnabled),
-							resource.TestCheckResourceAttr(resourceName, "state", testcase.expectedState),
+							resource.TestCheckResourceAttr(resourceName, "state", string(testcase.expectedState)),
 							testAccCheckRuleEnabled(ctx, resourceName, testcase.expectedState),
 						),
 					},
@@ -733,20 +858,20 @@ func TestAccEventsRule_migrateV0_Equivalent(t *testing.T) {
 		enabled           bool
 		state             string
 		expectedIsEnabled string
-		expectedState     string
+		expectedState     types.RuleState
 	}{
 		"enabled": {
 			enabled:           true,
-			state:             eventbridge.RuleStateEnabled,
+			state:             string(types.RuleStateEnabled),
 			expectedIsEnabled: "true",
-			expectedState:     eventbridge.RuleStateEnabled,
+			expectedState:     types.RuleStateEnabled,
 		},
 
 		"disabled": {
 			enabled:           false,
-			state:             eventbridge.RuleStateDisabled,
+			state:             string(types.RuleStateDisabled),
 			expectedIsEnabled: "false",
-			expectedState:     eventbridge.RuleStateDisabled,
+			expectedState:     types.RuleStateDisabled,
 		},
 	}
 
@@ -787,7 +912,7 @@ func TestAccEventsRule_migrateV0_Equivalent(t *testing.T) {
 						},
 						Check: resource.ComposeAggregateTestCheckFunc(
 							resource.TestCheckResourceAttr(resourceName, "is_enabled", testcase.expectedIsEnabled),
-							resource.TestCheckResourceAttr(resourceName, "state", testcase.expectedState),
+							resource.TestCheckResourceAttr(resourceName, "state", string(testcase.expectedState)),
 							testAccCheckRuleEnabled(ctx, resourceName, testcase.expectedState),
 						),
 					},
@@ -804,19 +929,9 @@ func testAccCheckRuleExists(ctx context.Context, n string, v *eventbridge.Descri
 			return fmt.Errorf("Not found: %s", n)
 		}
 
-		if rs.Primary.ID == "" {
-			return fmt.Errorf("No EventBridge Rule ID is set")
-		}
+		conn := acctest.Provider.Meta().(*conns.AWSClient).EventsClient(ctx)
 
-		eventBusName, ruleName, err := tfevents.RuleParseResourceID(rs.Primary.ID)
-
-		if err != nil {
-			return err
-		}
-
-		conn := acctest.Provider.Meta().(*conns.AWSClient).EventsConn(ctx)
-
-		output, err := tfevents.FindRuleByTwoPartKey(ctx, conn, eventBusName, ruleName)
+		output, err := tfevents.FindRuleByTwoPartKey(ctx, conn, rs.Primary.Attributes["event_bus_name"], rs.Primary.Attributes["name"])
 
 		if err != nil {
 			return err
@@ -828,28 +943,22 @@ func testAccCheckRuleExists(ctx context.Context, n string, v *eventbridge.Descri
 	}
 }
 
-func testAccCheckRuleEnabled(ctx context.Context, n string, want string) resource.TestCheckFunc {
+func testAccCheckRuleEnabled(ctx context.Context, n string, want types.RuleState) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[n]
 		if !ok {
 			return fmt.Errorf("Not found: %s", n)
 		}
 
-		eventBusName, ruleName, err := tfevents.RuleParseResourceID(rs.Primary.ID)
+		conn := acctest.Provider.Meta().(*conns.AWSClient).EventsClient(ctx)
+
+		output, err := tfevents.FindRuleByTwoPartKey(ctx, conn, rs.Primary.Attributes["event_bus_name"], rs.Primary.Attributes["name"])
 
 		if err != nil {
 			return err
 		}
 
-		conn := acctest.Provider.Meta().(*conns.AWSClient).EventsConn(ctx)
-
-		output, err := tfevents.FindRuleByTwoPartKey(ctx, conn, eventBusName, ruleName)
-
-		if err != nil {
-			return err
-		}
-
-		if got := aws.StringValue(output.State); got != want {
+		if got := output.State; got != want {
 			return fmt.Errorf("EventBridge Rule State = %v, want %v", got, want)
 		}
 
@@ -859,20 +968,14 @@ func testAccCheckRuleEnabled(ctx context.Context, n string, want string) resourc
 
 func testAccCheckRuleDestroy(ctx context.Context) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
-		conn := acctest.Provider.Meta().(*conns.AWSClient).EventsConn(ctx)
+		conn := acctest.Provider.Meta().(*conns.AWSClient).EventsClient(ctx)
 
 		for _, rs := range s.RootModule().Resources {
 			if rs.Type != "aws_cloudwatch_event_rule" {
 				continue
 			}
 
-			eventBusName, ruleName, err := tfevents.RuleParseResourceID(rs.Primary.ID)
-
-			if err != nil {
-				return err
-			}
-
-			_, err = tfevents.FindRuleByTwoPartKey(ctx, conn, eventBusName, ruleName)
+			_, err := tfevents.FindRuleByTwoPartKey(ctx, conn, rs.Primary.Attributes["event_bus_name"], rs.Primary.Attributes["name"])
 
 			if tfresource.NotFound(err) {
 				continue
@@ -891,7 +994,7 @@ func testAccCheckRuleDestroy(ctx context.Context) resource.TestCheckFunc {
 
 func testAccCheckRuleRecreated(i, j *eventbridge.DescribeRuleOutput) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
-		if aws.StringValue(i.Arn) == aws.StringValue(j.Arn) {
+		if aws.ToString(i.Arn) == aws.ToString(j.Arn) {
 			return fmt.Errorf("EventBridge rule not recreated, but expected it to be")
 		}
 		return nil
@@ -900,7 +1003,7 @@ func testAccCheckRuleRecreated(i, j *eventbridge.DescribeRuleOutput) resource.Te
 
 func testAccCheckRuleNotRecreated(i, j *eventbridge.DescribeRuleOutput) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
-		if aws.StringValue(i.Arn) != aws.StringValue(j.Arn) {
+		if aws.ToString(i.Arn) != aws.ToString(j.Arn) {
 			return fmt.Errorf("EventBridge rule recreated, but expected it to not be")
 		}
 		return nil
